refactor: rename autobump util to utils
