+++ src/commands/autobump-threads/list-threads.ts
@@ -3,7 +3,7 @@ import { SlashCommandSubcommandBuilder } from 'discord.js';
 import { Result } from 'oxide.ts';
 import { logger } from '../../utils/logger';
 import type { CommandHandler, Subcommand } from '../builder';
-import { listThreadsByGuild } from './util';
+import { listThreadsByGuild } from './utils';
 
 const data = new SlashCommandSubcommandBuilder().setName('list').setDescription('Show list of autobump threads');
